+++ lib/api/src/sys/store.rs
@@ -1,9 +1,8 @@
-use crate::sys::engineref::{AsEngineRef, EngineRef};
 use crate::sys::tunables::BaseTunables;
 use std::fmt;
 use std::sync::{Arc, RwLock};
 #[cfg(feature = "compiler")]
-use wasmer_compiler::{Engine, EngineBuilder, Tunables};
+use wasmer_compiler::{AsEngineRef, Engine, EngineBuilder, EngineRef, Tunables};
 use wasmer_vm::{init_traps, TrapHandler, TrapHandlerFn};
 
 use wasmer_vm::StoreObjects;
@@ -15,8 +14,6 @@ pub(crate) struct StoreInner {
     pub(crate) objects: StoreObjects,
     #[cfg(feature = "compiler")]
     pub(crate) engine: Engine,
-    #[cfg(feature = "compiler")]
-    pub(crate) tunables: Box<dyn Tunables + Send + Sync>,
     pub(crate) trap_handler: Option<Box<TrapHandlerFn<'static>>>,
 }
 
@@ -26,10 +23,7 @@ pub(crate) struct StoreInner {
 /// have been allocated during the lifetime of the abstract machine.
 ///
 /// The `Store` holds the engine (that is —amongst many things— used to compile
-/// the Wasm bytes into a valid module artifact), in addition to the
-#[cfg_attr(feature = "compiler", doc = "[`Tunables`]")]
-#[cfg_attr(not(feature = "compiler"), doc = "`Tunables`")]
-/// (that are used to create the memories, tables and globals).
+/// the Wasm bytes into a valid module artifact).
 ///
 /// Spec: <https://webassembly.github.io/spec/core/exec/runtime.html#store>
 pub struct Store {
@@ -44,8 +38,20 @@ impl Store {
     /// Creates a new `Store` with a specific [`Engine`].
     pub fn new(engine: impl Into<Engine>) -> Self {
         let engine = engine.into();
-        let target = engine.target().clone();
-        Self::new_with_tunables(engine, BaseTunables::for_target(&target))
+
+        // Make sure the signal handlers are installed.
+        // This is required for handling traps.
+        init_traps();
+
+        Self {
+            inner: Box::new(StoreInner {
+                objects: Default::default(),
+                engine: engine.cloned(),
+                trap_handler: None,
+            }),
+            engine: engine.cloned(),
+            trap_handler: Arc::new(RwLock::new(None)),
+        }
     }
 
     #[cfg(feature = "compiler")]
@@ -69,28 +75,16 @@ impl Store {
         engine: impl Into<Engine>,
         tunables: impl Tunables + Send + Sync + 'static,
     ) -> Self {
-        let engine = engine.into();
+        let mut engine = engine.into();
+        engine.set_tunables(tunables);
 
-        // Make sure the signal handlers are installed.
-        // This is required for handling traps.
-        init_traps();
-
-        Self {
-            inner: Box::new(StoreInner {
-                objects: Default::default(),
-                engine: engine.cloned(),
-                tunables: Box::new(tunables),
-                trap_handler: None,
-            }),
-            engine: engine.cloned(),
-            trap_handler: Arc::new(RwLock::new(None)),
-        }
+        Self::new(engine)
     }
 
     #[cfg(feature = "compiler")]
     /// Returns the [`Tunables`].
     pub fn tunables(&self) -> &dyn Tunables {
-        self.inner.tunables.as_ref()
+        self.engine.tunables()
     }
 
     #[cfg(feature = "compiler")]
@@ -205,37 +199,25 @@ impl AsStoreMut for Store {
 
 impl AsEngineRef for Store {
     fn as_engine_ref(&self) -> EngineRef<'_> {
-        EngineRef {
-            inner: &self.engine,
-            tunables: self.inner.tunables.as_ref(),
-        }
+        EngineRef::new(&self.engine)
     }
 }
 
 impl AsEngineRef for &Store {
     fn as_engine_ref(&self) -> EngineRef<'_> {
-        EngineRef {
-            inner: &self.engine,
-            tunables: self.inner.tunables.as_ref(),
-        }
+        EngineRef::new(&self.engine)
     }
 }
 
 impl AsEngineRef for StoreRef<'_> {
     fn as_engine_ref(&self) -> EngineRef<'_> {
-        EngineRef {
-            inner: &self.inner.engine,
-            tunables: self.inner.tunables.as_ref(),
-        }
+        EngineRef::new(&self.inner.engine)
     }
 }
 
 impl AsEngineRef for StoreMut<'_> {
     fn as_engine_ref(&self) -> EngineRef<'_> {
-        EngineRef {
-            inner: &self.inner.engine,
-            tunables: self.inner.tunables.as_ref(),
-        }
+        EngineRef::new(&self.inner.engine)
     }
 }
 
@@ -258,7 +240,7 @@ impl<'a> StoreRef<'a> {
     #[cfg(feature = "compiler")]
     /// Returns the [`Tunables`].
     pub fn tunables(&self) -> &dyn Tunables {
-        self.inner.tunables.as_ref()
+        self.inner.engine.tunables()
     }
 
     #[cfg(feature = "compiler")]
@@ -294,7 +276,7 @@ impl<'a> StoreMut<'a> {
     /// Returns the [`Tunables`].
     #[cfg(feature = "compiler")]
     pub fn tunables(&self) -> &dyn Tunables {
-        self.inner.tunables.as_ref()
+        self.inner.engine.tunables()
     }
 
     /// Returns the [`Engine`].
@@ -313,7 +295,7 @@ impl<'a> StoreMut<'a> {
 
     #[cfg(feature = "compiler")]
     pub(crate) fn tunables_and_objects_mut(&mut self) -> (&dyn Tunables, &mut StoreObjects) {
-        (self.inner.tunables.as_ref(), &mut self.inner.objects)
+        (self.inner.engine.tunables(), &mut self.inner.objects)
     }
 
     pub(crate) fn as_raw(&self) -> *mut StoreInner {
