+++ lib/compiler/src/engine/engineref.rs
@@ -1,28 +1,25 @@
+use super::Engine;
 use crate::Tunables;
-use wasmer_compiler::Engine;
 
-/// A temporary handle to an [`Engine`] and [`Tunables`].
+/// A temporary handle to an [`Engine`]
 /// EngineRef can be used to build a [`Module`][crate::sys::Module]
-/// It can be created directly with an [`Engine`] and [`Tunables`]
+/// It can be created directly with an [`Engine`]
 /// Or from anything implementing [`AsEngineRef`]
 /// like from [`Store`][crate::sys::Store] typicaly
 /// # Example
 ///
 /// ```
-/// # use wasmer::{Cranelift, EngineBuilder, BaseTunables, EngineRef, Module};
+/// # use crate::{Cranelift, EngineBuilder, BaseTunables, EngineRef, Module};
 /// #
 /// # let compiler = Cranelift::default();
 /// # let engine = EngineBuilder::new(compiler).engine();
-/// # let tunables = BaseTunables::for_target(&engine.target());
-/// # let engineref = EngineRef::new(&engine, &tunables);
+/// # let engineref = EngineRef::new(&engine);
 ///
 /// let module = Module::from_file(&engineref, "path/to/foo.wasm");
 /// ```
 pub struct EngineRef<'a> {
     /// The inner engine
     pub(crate) inner: &'a Engine,
-    /// optionnal tunnables
-    pub(crate) tunables: &'a dyn Tunables,
 }
 
 impl<'a> EngineRef<'a> {
@@ -32,14 +29,11 @@ impl<'a> EngineRef<'a> {
     }
     /// Get the [`Tunables`]
     pub fn tunables(&self) -> &dyn Tunables {
-        self.tunables
+        self.inner.tunables()
     }
     /// Create an EngineRef from an Engine and Tunables
-    pub fn new(engine: &'a Engine, tunables: &'a dyn Tunables) -> Self {
-        EngineRef {
-            inner: engine,
-            tunables,
-        }
+    pub fn new(engine: &'a Engine) -> Self {
+        EngineRef { inner: engine }
     }
 }
 
@@ -51,9 +45,6 @@ pub trait AsEngineRef {
 
 impl AsEngineRef for EngineRef<'_> {
     fn as_engine_ref(&self) -> EngineRef<'_> {
-        EngineRef {
-            inner: self.inner,
-            tunables: self.tunables,
-        }
+        EngineRef { inner: self.inner }
     }
 }
