+++ lib/compiler/src/engine/inner.rs
@@ -4,7 +4,11 @@ use crate::engine::builder::EngineBuilder;
 #[cfg(not(target_arch = "wasm32"))]
 use crate::Artifact;
 #[cfg(not(target_arch = "wasm32"))]
+use crate::BaseTunables;
+#[cfg(not(target_arch = "wasm32"))]
 use crate::CodeMemory;
+#[cfg(not(target_arch = "wasm32"))]
+use crate::{AsEngineRef, EngineRef};
 #[cfg(feature = "compiler")]
 use crate::{Compiler, CompilerConfig};
 #[cfg(not(target_arch = "wasm32"))]
@@ -36,6 +40,8 @@ pub struct Engine {
     /// The target for the compiler
     target: Arc<Target>,
     engine_id: EngineId,
+    #[cfg(not(target_arch = "wasm32"))]
+    tunables: Arc<dyn Tunables + Send + Sync>,
 }
 
 impl Engine {
@@ -46,6 +52,8 @@ impl Engine {
         target: Target,
         features: Features,
     ) -> Self {
+        #[cfg(not(target_arch = "wasm32"))]
+        let tunables = BaseTunables::for_target(&target);
         Self {
             inner: Arc::new(Mutex::new(EngineInner {
                 compiler: Some(compiler_config.compiler()),
@@ -57,6 +65,8 @@ impl Engine {
             })),
             target: Arc::new(target),
             engine_id: EngineId::default(),
+            #[cfg(not(target_arch = "wasm32"))]
+            tunables: Arc::new(tunables),
         }
     }
 
@@ -74,6 +84,9 @@ impl Engine {
     /// Headless engines can't compile or validate any modules,
     /// they just take already processed Modules (via `Module::serialize`).
     pub fn headless() -> Self {
+        let target = Target::default();
+        #[cfg(not(target_arch = "wasm32"))]
+        let tunables = BaseTunables::for_target(&target);
         Self {
             inner: Arc::new(Mutex::new(EngineInner {
                 #[cfg(feature = "compiler")]
@@ -85,8 +98,10 @@ impl Engine {
                 #[cfg(not(target_arch = "wasm32"))]
                 signatures: SignatureRegistry::new(),
             })),
-            target: Arc::new(Target::default()),
+            target: Arc::new(target),
             engine_id: EngineId::default(),
+            #[cfg(not(target_arch = "wasm32"))]
+            tunables: Arc::new(tunables),
         }
     }
 
@@ -128,12 +143,12 @@ impl Engine {
     /// Compile a WebAssembly binary
     #[cfg(feature = "compiler")]
     #[cfg(not(target_arch = "wasm32"))]
-    pub fn compile(
-        &self,
-        binary: &[u8],
-        tunables: &dyn Tunables,
-    ) -> Result<Arc<Artifact>, CompileError> {
-        Ok(Arc::new(Artifact::new(self, binary, tunables)?))
+    pub fn compile(&self, binary: &[u8]) -> Result<Arc<Artifact>, CompileError> {
+        Ok(Arc::new(Artifact::new(
+            self,
+            binary,
+            self.tunables.as_ref(),
+        )?))
     }
 
     /// Compile a WebAssembly binary
@@ -186,6 +201,24 @@ impl Engine {
     /// Clone the engine
     pub fn cloned(&self) -> Self {
         self.clone()
+    }
+
+    /// Attach a Tunable to this engine
+    #[cfg(not(target_arch = "wasm32"))]
+    pub fn set_tunables(&mut self, tunables: impl Tunables + Send + Sync + 'static) {
+        self.tunables = Arc::new(tunables);
+    }
+
+    /// Get a reference to attached Tunable of this engine
+    #[cfg(not(target_arch = "wasm32"))]
+    pub fn tunables(&self) -> &dyn Tunables {
+        self.tunables.as_ref()
+    }
+}
+
+impl AsEngineRef for Engine {
+    fn as_engine_ref(&self) -> EngineRef {
+        EngineRef { inner: self }
     }
 }
