+++ lib/compiler/src/engine/mod.rs
@@ -1,5 +1,8 @@
 //! The Wasmer Engine.
 
+#[cfg(feature = "translator")]
+#[cfg(not(target_arch = "wasm32"))]
+mod engineref;
 mod error;
 #[cfg(not(target_arch = "wasm32"))]
 mod resolver;
@@ -25,13 +28,16 @@ mod link;
 #[cfg(not(target_arch = "wasm32"))]
 mod unwind;
 
+#[cfg(feature = "translator")]
+#[cfg(not(target_arch = "wasm32"))]
+pub use self::engineref::{AsEngineRef, EngineRef};
 pub use self::error::{InstantiationError, LinkError};
 #[cfg(not(target_arch = "wasm32"))]
 pub use self::resolver::resolve_imports;
 #[cfg(not(target_arch = "wasm32"))]
 pub use self::trap::*;
 #[cfg(not(target_arch = "wasm32"))]
-pub use self::tunables::Tunables;
+pub use self::tunables::{BaseTunables, Tunables};
 
 #[cfg(feature = "translator")]
 #[cfg(not(target_arch = "wasm32"))]
